+++ LibC/include/stdio.h
@@ -27,6 +27,26 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 extern "C" {
 #endif
 
+#define BUFSIZ 1024
+
+#ifndef EOF
+#define EOF -1
+#endif
+
+#define FILENAME_MAX 256
+#define FOPEN_MAX 8
+#define L_tmpnam 256
+
+#ifndef NULL
+#define NULL ((void*)0)
+#endif
+
+#define TMP_MAX 256
+
+#define SEEK_SET 0
+#define SEEK_CUR 1
+#define SEEK_END 2
+
 struct FILE {
     fd_t descriptor;
     unsigned long flags;
@@ -44,44 +64,69 @@ extern FILE* stddebug;
 #define stderr stderr
 #define stddebug stddebug
 
-#define SEEK_SET 0
+typedef long int fpos_t;
 
-#define FOPEN_MAX 8
+#ifndef size_t
+typedef unsigned long int size_t;
+#endif
 
-int getc();
-int fgetc(FILE* file);
+int remove(const char* filename);
+int rename(const char* oldname, const char* newname);
+FILE* tmpfile(void);
+char* tmpnam(char* str);
 
-void putc(const char c);
-inline void putchar(const char c) { putc(c); }
-void fputc(FILE* file, const char c);
-void dbgputc(const char c);
-
-void puts(const char* str);
-void fputs(FILE* file, const char* str);
-void dbgputs(const char* str);
-
-int printf(const char* format, ...);
-int vprintf(const char* format, va_list args);
-int fprintf(FILE* file, const char* format, ...);
-int vfprintf(FILE* file, const char* format, va_list args);
-int sprintf(char* str, const char* format, ...);
-
-int dbgprintf(const char* format, ...);
-int dbgvprintf(const char* format, va_list args);
-
-size_t fwrite(const void* ptr, const size_t size, const size_t count, FILE* file);
-size_t fread(void* ptr, size_t size, size_t count, FILE* stream);
-
-FILE* fopen(const char* filename, const char* mode);
 int fclose(FILE* stream);
-
 int fflush(FILE* stream);
+FILE* fopen(const char* filename, const char* mode);
+FILE* freopen(const char* filename, const char* mode, FILE* stream);
+void setbuf(FILE* stream, char* buffer);
+int setvbuf(FILE* stream, char* buffer, int mode, size_t size);
 
+int fprintf(FILE* stream, const char* format, ...);
+int fscanf(FILE* stream, const char* format, ...);
+int printf(const char* format, ...);
+int scanf(const char* format, ...);
+int snprintf(char* str, size_t size, const char* format, ...);
+int sprintf(char* str, const char* format, ...);
+int sscanf(const char* str, const char* format, ...);
+int vfprintf(FILE* stream, const char* format, va_list arg);
+int vfscanf(FILE* stream, const char* format, va_list arg);
+int vprintf(const char* format, va_list arg);
+int vscanf(const char* format, va_list arg);
+int vsnprintf(char* str, size_t size, const char* format, va_list arg);
+int vsprintf(char* str, const char* format, va_list arg);
+int vsscanf(const char* str, const char* format, va_list arg);
+
+int fgetc(FILE* stream);
+char* fgets(char* str, int num, FILE* stream);
+int fputc(int character, FILE* stream);
+int fputs(const char* str, FILE* stream);
+int getc(FILE* stream);
+int getchar(void);
+char* gets(char* str);
+int putc(int character, FILE* stream);
+int putchar(int character);
+int puts(const char* str);
+int ungetc(int character, FILE* stream);
+
+size_t fread(void* ptr, size_t size, size_t count, FILE* stream);
+size_t fwrite(const void* ptr, size_t size, size_t count, FILE* stream);
+
+int fgetpos(FILE* stream, fpos_t* pos);
 int fseek(FILE* stream, long int offset, int origin);
+int fsetpos(FILE* stream, const fpos_t* pos);
+long int ftell(FILE* stream);
 void rewind(FILE* stream);
 
-long int ftell(FILE* stream);
-void setbuf(FILE* stream, char* buffer);
+void clearerr(FILE* stream);
+int feof(FILE* stream);
+int ferror(FILE* stream);
+void perror(const char* str);
+
+int dbgputc(const char c);
+int dbgputs(const char* str);
+int dbgprintf(const char* format, ...);
+int dbgvprintf(const char* format, va_list args);
 
 #ifdef __cplusplus
 }