+++ LibC/src/string.c
@@ -15,7 +15,8 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "string.h"
+#include <string.h>
+#include <errno.h>
 
 size_t strlen(const char* str) {
     size_t len = 0;
@@ -149,4 +150,13 @@ char* strrchr(const char* str, int character) {
     if (found)
         return &(((char*)str)[last]);
     return NULL;
+}
+
+char* strerror(int errnum) {
+    #define _STRERROR_CASE(i, n) case i: return n;
+    switch (errnum) {
+        ENUMERATE_ERRNO_CODES(_STRERROR_CASE)
+    }
+    #undef _STRERROR_CASE
+    return "Unknown error";
 }