+++ LibC/src/util.c
@@ -17,6 +17,9 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "util.h"
 
+// x86_64 has architecture specific optimised implementations of memset and memcpy.
+#ifndef __x86_64__
+
 void* memset(void* dst, const uint8_t value, const size_t n) {
     uint8_t* d = (uint8_t*) dst;
 
@@ -37,6 +40,8 @@ void* memcpy(void* dst, const void* src, const size_t n) {
 
     return dst;
 }
+
+#endif
 
 void* memmove(void* dst, const void* src, const size_t n) {
     // OK, since we know that memcpy copies forwards