+++ Programs/src/echo.cpp
@@ -20,6 +20,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 int main(int argc, char** argv) {
     for (int i = 1; i < argc; i++)
         printf("%s%s", argv[i], i < (argc - 1) ? " " : "");
-    putc('\n');
+    putchar('\n');
     return 0;
 }